The cake is a lie
The sheet is a lie. The sheet is a lie. The sheet is a lie. Do not look
at the sheet. Do not trust the sheet. The sheet is a lie.

(program's parameter is an idList rather than a simple id. This meant
that every file to do with the parser got modified)