The loader works
!!!!!

Yes I know the : for nt's outside of code; without nt on the inside is
strange, but it's the simplest way to ENSURE the grammar is unambiguous
for parsing my decorated grammar.

[otherwise, I need a lot of look ahead to determine the context of the
nonterminal being used. Is the NT for code, or is it being used as a
nonterminal to start productions? HARD TO SAY!]